Move simd library closer to hash join (#25415)
