+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/AVX2_algo/avx2_algo.cpp
@@ -1,0 +1,9 @@
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/algo.h>
+#include <immintrin.h>
+#include <avxintrin.h>
+
+template<>
+THolder<Perfomancer::Interface> Perfomancer::Create<AVX2Trait>() {
+    Cerr << "AVXTrait ";
+    return MakeHolder<Algo<AVX2Trait>>();
+}
