+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/AVX2_algo/ya.make
@@ -1,0 +1,11 @@
+IF (ARCH_X86_64 AND OS_LINUX)
+
+LIBRARY()
+
+CFLAGS(-mavx2)
+
+SRCS(avx2_algo.cpp)
+
+END()
+
+ENDIF()
