+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/Fallback_algo/fallback_algo.cpp
@@ -1,0 +1,7 @@
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/algo.h>
+
+template<>
+THolder<Perfomancer::Interface> Perfomancer::Create<FallbackTrait>() {
+    Cerr << "FallbackTrait ";
+    return MakeHolder<Algo<FallbackTrait>>();
+}
