+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/Fallback_algo/ya.make
@@ -1,0 +1,9 @@
+IF (ARCH_X86_64 AND OS_LINUX)
+
+LIBRARY()
+
+SRCS(fallback_algo.cpp)
+
+END()
+
+ENDIF()
