+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/SSE42_algo/sse42_algo.cpp
@@ -1,0 +1,8 @@
+#include <ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/algo.h>
+#include <immintrin.h>
+
+template<>
+THolder<Perfomancer::Interface> Perfomancer::Create<SSE42Trait>() {
+    Cerr << "SSETrait ";
+    return MakeHolder<Algo<SSE42Trait>>();
+}
