+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/exec/runtime_dispatching/SSE42_algo/ya.make
@@ -1,0 +1,11 @@
+IF (ARCH_X86_64 AND OS_LINUX)
+
+LIBRARY()
+
+CFLAGS(-msse4.2)
+
+SRCS(sse42_algo.cpp)
+
+END()
+
+ENDIF()
