+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/ut/ya.make
@@ -1,0 +1,11 @@
+IF (ARCH_X86_64 AND OS_LINUX)
+
+UNITTEST_FOR(ydb/library/yql/dq/comp_nodes/hash_join_utils/simd)
+
+SRCS(
+    simd_ut.cpp
+)
+
+END()
+
+ENDIF()
