+++ ydb/library/yql/dq/comp_nodes/hash_join_utils/simd/ya.make
@@ -1,0 +1,14 @@
+IF (ARCH_X86_64 AND OS_LINUX)
+
+LIBRARY()
+
+PEERDIR(library/cpp/digest/crc32c)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+    exec
+)
+
+ENDIF()
