--- ydb/library/yql/utils/simd/exec/add_columns/ya.make
@@ -1,11 +1,0 @@
-PROGRAM(add_columns)
-
-SRCS(main.cpp)
-
-SIZE(MEDIUM)
-
-CFLAGS(-mavx2)
-
-PEERDIR(ydb/library/yql/utils/simd)
-
-END()
