--- ydb/library/yql/utils/simd/exec/pack_tuple/ya.make
@@ -1,9 +1,0 @@
-PROGRAM()
-
-SRCS(main.cpp)
-
-CFLAGS(-mavx2)
-
-PEERDIR(ydb/library/yql/utils/simd)
-
-END()
