--- ydb/library/yql/utils/simd/exec/runtime_dispatching/AVX2_algo/avx2_algo.cpp
@@ -1,9 +1,0 @@
-#include <ydb/library/yql/utils/simd/exec/runtime_dispatching/algo.h>
-#include <immintrin.h>
-#include <avxintrin.h>
-
-template<>
-THolder<Perfomancer::Interface> Perfomancer::Create<AVX2Trait>() {
-    Cerr << "AVXTrait ";
-    return MakeHolder<Algo<AVX2Trait>>();
-}
