--- ydb/library/yql/utils/simd/exec/runtime_dispatching/AVX2_algo/ya.make
@@ -1,7 +1,0 @@
-LIBRARY()
-
-CFLAGS(-mavx2)
-
-SRCS(avx2_algo.cpp)
-
-END()
