--- ydb/library/yql/utils/simd/exec/runtime_dispatching/Fallback_algo/fallback_algo.cpp
@@ -1,7 +1,0 @@
-#include <ydb/library/yql/utils/simd/exec/runtime_dispatching/algo.h>
-
-template<>
-THolder<Perfomancer::Interface> Perfomancer::Create<FallbackTrait>() {
-    Cerr << "FallbackTrait ";
-    return MakeHolder<Algo<FallbackTrait>>();
-}
