--- ydb/library/yql/utils/simd/exec/runtime_dispatching/Fallback_algo/ya.make
@@ -1,5 +1,0 @@
-LIBRARY()
-
-SRCS(fallback_algo.cpp)
-
-END()
