--- ydb/library/yql/utils/simd/exec/runtime_dispatching/SSE42_algo/sse42_algo.cpp
@@ -1,8 +1,0 @@
-#include <ydb/library/yql/utils/simd/exec/runtime_dispatching/algo.h>
-#include <immintrin.h>
-
-template<>
-THolder<Perfomancer::Interface> Perfomancer::Create<SSE42Trait>() {
-    Cerr << "SSETrait ";
-    return MakeHolder<Algo<SSE42Trait>>();
-}
