--- ydb/library/yql/utils/simd/exec/runtime_dispatching/SSE42_algo/ya.make
@@ -1,7 +1,0 @@
-LIBRARY()
-
-CFLAGS(-msse4.2)
-
-SRCS(sse42_algo.cpp)
-
-END()
