--- ydb/library/yql/utils/simd/exec/tuples_to_bucket/ya.make
@@ -1,7 +1,0 @@
-PROGRAM()
-
-SRCS(main.cpp)
-
-PEERDIR(ydb/library/yql/utils/simd)
-
-END()
