--- ydb/library/yql/utils/simd/ut/ya.make
@@ -1,7 +1,0 @@
-UNITTEST_FOR(ydb/library/yql/utils/simd)
-
-SRCS(
-    simd_ut.cpp
-)
-
-END()
