--- ydb/library/yql/utils/simd/ya.make
@@ -1,10 +1,0 @@
-LIBRARY()
-
-PEERDIR(library/cpp/digest/crc32c)
-
-END()
-
-RECURSE_FOR_TESTS(
-    ut
-    exec
-)
